Pre-merge clean up of code and project
