+++ Assets/Scripts/Weather System/Controllers/Custom Controllers/FogController.cs
