+++ Assets/Scripts/Weather System/Director.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a111f57741163304db8c0e5712294cb2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
