+++ Assets/Scripts/Weather System/Director/WeatherState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 659327940e114a04ea74e701d81646e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
